# docs/swagger.yaml
openapi: 3.0.0
servers:
  - url: "https://api.eesast.com"
    description: Remote API Server
  - url: "http://localhost:{port}"
    description: Local API Server
    variables:
      port:
        description: 端口号
        default: "28888"
info:
  title: API Backend for EESAST
  description: |
    清华大学电子工程系学生科协后端 API。了解更多 EESAST 内容，请访问：[GitHub 主页](https://github.com/eesast/)或 [EESAST 官方网站](https://eesast.com/)。
  version: 1.0.0
  contact:
    name: API Support
    email: i@raineggplant.com
...
paths:
...
  "/v1/codes":
    get:
      tags:
        - codes
      description: 需要已登录用户
      operationId: getCodes
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: contestId
          description: 所属比赛
          required: false
          schema:
            type: number
        - in: query
          name: teamId
          description: 队伍Id
          required: false
          schema:
            type: number
        - name: begin
          in: query
          description: 按时间降序排列，起始代码的位置
          schema:
            type: number
        - name: end
          in: query
          description: 按时间降序排列，结束代码的位置
          schema:
            type: number
      responses:
        200:
          description: Successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Code"
        403:
          description: 403 Forbidden
        404:
          description: 404 Not Found
        422:
          description: 422 Missing contents
        500:
          description: Server Error

    post:
      description: 上传代码
      tags:
        - codes
      operationId: addCode
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/addCode"
      responses:
        201:
          description: Successful operation
          headers:
            Location:
              description: 新上传代码的 URL
              schema:
                type: string
        422:
          description: 422 Missing contents
        500:
          description: Server Error
  "/v1/codes/{codeId}":
    get:
      tags:
        - codes
      description: 通过ID获取代码
      summary: Describe this route
      operationId: getCodeById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: codeId
          required: true
          description: 代码Id
          schema:
            type: number
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Code"
        403:
          description: Forbidden
        500:
          description: Server Error

    put:
      tags:
        - codes
      summary: 更新已存在的代码
      description: 需要 `root` `organizer` 角色，所属队伍成员
      operationId: updateCode
      security:
        - bearerAuth: []
      parameters:
        - name: codeId
          in: path
          description: 待更新代码的 ID
          required: true
          schema:
            type: number
      requestBody:
        $ref: "#/components/requestBodies/updateCode"
      responses:
        204:
          description: Successful operation
          headers:
            Location:
              description: 被更新代码的 URL
              schema:
                type: string
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: Team / Code does not exist"
        500:
          description: Server error

    delete:
      tags:
        - codes
      summary: 删除代码
      description: 需要 `root` 角色，所属队伍成员
      operationId: deleteCode
      security:
        - bearerAuth: []
      parameters:
        - name: codeId
          in: path
          description: 要删除代码的 ID
          required: true
          schema:
            type: number
      responses:
        204:
          description: Successful operation
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: Code / Team does not exist"
        500:
          description: Server error
  "/v1/codes/{codeId}/compile":
    post:
      description: 编译代码
      tags:
        - codes
      operationId: compileCode
      security:
        - bearerAuth: []
      parameters:
        - name: codeId
          in: path
          description: 要编译代码的 ID
          required: true
          schema:
            type: number
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  type: number
      responses:
        200:
          description: Successful operation
          headers:
            Location:
              description: 编译信息（暂未区分是否成功）
              schema:
                type: string
        400:
          description: 400 Error (TODO Add info)
        404:
          description: 404 Code does not exist
        409:
          description: 409 Code is compiling
        500:
          description: Server Error

    put:
      tags:
        - codes
      summary: 更新编译信息
      description: 由server发起
      operationId: updateCodeCompileInfo
      security:
        - bearerAuth: []
      parameters:
        - name: codeId
          in: path
          description: 代码 ID
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - compileInfo
              properties:
                status:
                  type: string
      responses:
        401:
          $ref: "#/components/responses/UnauthorizedError"
        403:
          description: "Forbidden: Permission denied"
        404:
          description: "Not Found: Code does not exist / Docker container does not exist"
        422:
          description: "Unprocessable Entity: Missing form data"
        500:
          description: Server error
        503:
          description: "Service Unavailable: Failed to stop or remove docker container"
components:
  schemas:
...
    Date:
      type: string
      format: date-time
      description: >-
        日期对象。使用时采用 ISO
        格式字符串表示。[更多信息](https://www.iso.org/iso-8601-date-and-time-format.html)
...
    Code:
      type: object
      required:
        - id
        - contestId
        - teamId
        - name
        - content
        - language
        - createdAt
        - createdBy
        - updatedAt
        - updatedBy
      properties:
        id:
          type: number
          description: 自增值
        contestId:
          type: number
          description: 对应比赛
        teamId:
          type: number
          description: 归属队伍
        name:
          type: string
          description: 文件名
        content:
          type: string
          description: 代码内容
        language:
          type: string
          description: 代码所用语言
        createdAt:
          $ref: "#/components/schemas/Date"
        createdBy:
          type: number
        updatedAt:
          $ref: "#/components/schemas/Date"
        updatedBy:
          type: number
  requestBodies:
...
    addCode:
      content:
        application/json:
          schema:
            type: object
            required:
              - name
              - teamId
              - contestId
              - content
              - language
            properties:
              name:
                type: string
              contestId:
                type: number
              teamId:
                type: number
              content:
                type: string
                description: 代码内容
              language:
                type: string
                description: 代码语言
      description: 上传代码的提交格式
      required: true
    updateCode:
      required: true
      content:
        application/json:
          schema:
            required:
              - name
              - content
              - language
            properties:
              name:
                type: string
              content:
                type: string
                description: 代码内容
              language:
                type: string
                description: 代码语言
  responses:
    BadRequest:
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                description: 错误列表
                items:
                  type: object
                  properties:
                    path:
                      type: string
                      description: 错误位置
                      example: ".params.articleId"
                    message:
                      type: string
                      description: 错误信息
                      example: "should be number"
                    errorCode:
                      type: string
                      description: 错误码
                      example: "type.openapi.validation"
      description: 包含请求错误信息的 JSON
    UnauthorizedError:
      description: "Access token is missing, invalid or expired / Permission denied"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT